% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageHeatmap.R
\name{AverageHeatmap}
\alias{AverageHeatmap}
\title{Plot averaged gene expression cross cluster cells}
\usage{
AverageHeatmap(
  object,
  markerGene,
  group.by = "ident",
  assays = "RNA",
  slot = "data",
  htCol = c("#0099CC", "white", "#CC0033"),
  colseed = 666,
  htRange = c(-2, 0, 2),
  annoCol = "none",
  annoColType = "light",
  annoColTypeAlpha = 0,
  row_title = "Cluster top Marker genes",
  row_names_side = "left",
  border = FALSE,
  fontsize = 10,
  column_names_rot = 45
)
}
\arguments{
\item{object}{object seurat object.}

\item{markerGene}{Your marker genes.}

\item{group.by}{Categories for grouping (e.g, ident, replicate, celltype). 'ident' by default.}

\item{assays}{Which assays to use. Default is "RNA" assays.}

\item{slot}{Slot(s) to use. Default is "data"}

\item{htCol}{Heatmap colors. Default is c("#0099CC", "white", "#CC0033").}

\item{colseed}{Cluster annotaion colors seed, these colors are produed randomly, so you can give a seed to assure produce same colors.  Default is 666.}

\item{htRange}{Heatmap values range. Default is c(-2, 0, 2).}

\item{annoCol}{You can specify your own annotation clusters colors vectors. Default is 'none'.}

\item{annoColType}{Cluster annotaion colors type (bright, light, dark and random). Default is light.}

\item{annoColTypeAlpha}{Cluster annotaion colors transparency. Default is 0.}

\item{row_title}{Heatmap row title. Default is "Cluster top Marker genes".}

\item{row_names_side}{Heatmap gene name side. Default is "left".}

\item{border}{Whether to shOw heatmap border. Default is "FALSE".}

\item{fontsize}{Heatmap gene name fontsize. Default is 10.}

\item{column_names_rot}{Cluster name rotation. Default is 45.}
}
\value{

}
\description{
Plot averaged gene expression cross cluster cells
}
\examples{
httest <- system.file("extdata", "htdata.RDS", package = "scRNAtoolVis")
pbmc <- readRDS(httest)

# load markergene
markergene <- system.file("extdata", "top5pbmc.markers.csv", package = "scRNAtoolVis")
markers <- read.table(markergene, sep = ',', header = TRUE)

# plot
AverageHeatmap(object = pbmc,
              markerGene = markers$gene)

# change color
AverageHeatmap(object = pbmc,
              markerGene = markers$gene,
              htCol = c("#339933", "#FFCC00", "#FF0033"))
}
