% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellRatioPlot.R
\name{cellRatioPlot}
\alias{cellRatioPlot}
\title{using cellRatioPlot to visualize cell percent ratio across clusters}
\arguments{
\item{object}{object seurat object.}

\item{sample.name}{sample name in meta info, default NULL.}

\item{celltype.name}{celltype name in meta info, default NULL.}

\item{col.width}{column width, default 0.7.}

\item{flow.alpha}{flow color alpha, default 0.25.}

\item{flow.curve}{flow curve rate, default 0.}

\item{fill.col}{supply your own fill colors, default NULL.}
}
\value{
a ggplot object.
}
\description{
using cellRatioPlot to visualize cell percent ratio across clusters
}
\author{
Junjun Lao
}
