% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCornerAxes.R
\name{clusterCornerAxes}
\alias{clusterCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE cluster figures}
\arguments{
\item{object}{seurat object.}

\item{reduction}{"string",reduction type (umap/tsne).}

\item{groupFacet}{"string",give the column name in seurat metadata to facet plot.}

\item{clusterCol}{"string",the point color to group by,cluster name,defaults "seurat_clusters".}

\item{pSize}{"num",point size.}

\item{noSplit}{'logic',whether to split/facet the plot,defaults "TRUE".}

\item{nrow}{"num",rows to plot when noSplit = FALSE.}

\item{relLength}{'num',the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num",the relative distance of corner axis label to axis.}

\item{axes}{"string",show multiple corner axis or only one (mul/one),defaults "mul".}

\item{legendPos}{"string",legend position same as ggplot theme function,defaults "right".}

\item{lineTextcol}{"string",corner line and label color,defaults "black".}

\item{stripCol}{"string",facet balckground color,defaults "white".}

\item{arrowType}{"string",arrow type (open/closed),defaults "closed".}
}
\value{

}
\description{
Add corner axes on seurat UMAP/tSNE cluster figures
}
\examples{
test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
clusterCornerAxes(object = tmp,reduction = 'umap',
                  noSplit = TRUE)

# arrowType
clusterCornerAxes(object = tmp,reduction = 'umap',
                  noSplit = TRUE,arrowType = 'open')

# facet by metadata column "orig.ident"
clusterCornerAxes(object = tmp,reduction = 'umap',
                  noSplit = FALSE,groupFacet = 'orig.ident',
                  relLength = 0.5)

# retain only one axes
clusterCornerAxes(object = tmp,reduction = 'umap',
                  noSplit = FALSE,groupFacet = 'orig.ident',
                  relLength = 0.5,
                  axes = 'one')

# line color
clusterCornerAxes(object = tmp,reduction = 'umap',
                  noSplit = FALSE,groupFacet = 'orig.ident',
                  relLength = 0.5,
                  lineTextcol = 'grey50')

# tsne
clusterCornerAxes(object = tmp,reduction = 'tsne',
                  noSplit = FALSE,groupFacet = 'orig.ident',
                  relLength = 0.5)
}
