% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLegend.R
\name{drawLegend}
\alias{drawLegend}
\title{using drawLegend to add cluster id in legend}
\arguments{
\item{object}{object seurat object, default NULL.}

\item{plot}{ggplot object, default NULL.}

\item{cellType}{cell type column name, default NULL.}

\item{clusters}{cluster id column name, default NULL.}

\item{ncol}{ncols to draw legend(1/2), default 1.}

\item{col}{point color, default hue_pal().}

\item{pt.size}{point size, default 8.}

\item{text.size}{text size, default 4.}
}
\value{
combine plot
}
\description{
using drawLegend to add cluster id in legend
}
\author{
Junjun Lao
}
