% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCornerAxes.R
\name{FeatureCornerAxes}
\alias{FeatureCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures}
\arguments{
\item{object}{object seurat object.}

\item{reduction}{"string",reduction type (umap/tsne).}

\item{features}{"string",the gene you want to plot.}

\item{groupFacet}{"string",give the column name in seurat metadata to facet plot.}

\item{relLength}{"num",the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num",the relative distance of corner axis label to axis.}

\item{aspect.ratio}{"num", plot width and height ratio, defaults NULL.}

\item{low}{"string",point color with low expression.}

\item{high}{"string",point color with high expression.}

\item{axes}{"string",show multiple corner axis or only one (mul/one),defaults "mul".}

\item{legendPos}{"string",legend position same as ggplot theme function,defaults "right".}

\item{RowSample}{"num",rows for groupFacet.}

\item{ColGene}{"num",cols for features.}

\item{stripCol}{"string",facet balckground color,defaults "white".}

\item{pSize}{"num",point size.}

\item{arrowType}{"string",arrow type (open/closed),defaults "closed".}

\item{lineTextcol}{"string",facet balckground color,defaults "white".}

\item{cornerTextSize}{"num", the corner label text size, defaults is 5.}

\item{base_size}{"num", theme base size, defaults is 14.}

\item{themebg}{Another theme style, defaults is 'default', or 'bwCorner'.}
}
\value{
Return a ggplot.
}
\description{
Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures
}
\examples{

test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
FeatureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)

# one axes
FeatureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  features = c("Actb", "Ythdc1", "Ythdf2"),
  relLength = 0.5, relDist = 0.2,
  axes = "one",
  lineTextcol = "grey50"
)

# tsne
FeatureCornerAxes(
  object = tmp, reduction = "tsne",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)


}
