% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurePlot.R
\name{featurePlot}
\alias{featurePlot}
\title{featurePlot Function}
\usage{
featurePlot(
  object = NULL,
  dim = "umap",
  genes = NULL,
  nrow = NULL,
  ncol = NULL,
  quantile.val = 1,
  color = NULL,
  rm.axis = FALSE,
  rm.legend = FALSE,
  add.rect = FALSE,
  add.corArrow = FALSE,
  add.strip = FALSE,
  corLabel.dist = 0.08,
  arrow.len = 0.2,
  arrow.label.size = 6,
  plot.size = 0.6,
  keep.oneCor = FALSE,
  xlab = NULL,
  ylab = NULL,
  respect = TRUE,
  point.size = 1
)
}
\arguments{
\item{object}{A Seurat object containing the data.}

\item{dim}{The dimension to use for plotting, default is "umap".}

\item{genes}{A character vector of gene names or feature names to be plotted.}

\item{nrow}{Number of rows in the plot grid.}

\item{ncol}{Number of columns in the plot grid.}

\item{quantile.val}{The quantile value to determine the color cutoff for each gene.}

\item{color}{A vector of colors to be used for plotting, defaults to a
predefined set of colors.}

\item{rm.axis}{Logical value indicating whether to remove axis labels and ticks,
defaults to FALSE.}

\item{rm.legend}{Logical value indicating whether to remove the color legend,
defaults to FALSE.}

\item{add.rect}{Logical value indicating whether to add a rectangle around
each plot panel, defaults to FALSE.}

\item{add.corArrow}{Logical value indicating whether to add arrows indicating
the correlation direction, defaults to FALSE.}

\item{add.strip}{Logical value indicating whether to add a strip at the top of
each plot panel, defaults to FALSE.}

\item{corLabel.dist}{Distance between the corner arrows and the axis labels.}

\item{arrow.len}{Length of the corner arrows.}

\item{arrow.label.size}{Font size of the corner arrow labels.}

\item{plot.size}{Size of each individual scatter plot.}

\item{keep.oneCor}{Logical value indicating whether to keep only one set of
corner arrows across the entire plot grid, defaults to FALSE.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{respect}{Logical value indicating whether to respect the specified number
of rows and columns in the plot grid, defaults to TRUE.}

\item{point.size}{the point size, default 1.}
}
\description{
This function creates a scatter plot for multiple genes or features from a
Seurat object.
}
\examples{
\dontrun{
# Assuming 'seurat_obj' is a Seurat object
featurePlot(object = seurat_obj, genes = c("gene1", "gene2", "gene3"), nrow = 2, ncol = 2)
}

}
\author{
Jun Zhang
}
