% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjDotPlot.R
\name{jjDotPlot}
\alias{jjDotPlot}
\title{using dotplot to visualize gene expression}
\arguments{
\item{object}{seurat object, default NULL.}

\item{assay}{the assay to be selected, default NULL.}

\item{slot}{gene expression data to be selected, default "data".}

\item{id}{the cell clusters id in the metadata info, default "seurat_clusters".}

\item{split.by}{the group name to split, default NULL.}

\item{split.by.aesGroup}{whether the dot color filled by group, default FALSE.}

\item{gene}{the genes to be drawed in plot, default NULL.}

\item{markerGene}{the marker genes with celltype info to be drawed, default NULL.}

\item{point.geom}{the ggplot "point" geom layer to be shown, default TRUE.}

\item{point.shape}{the point shape,default 21.}

\item{tile.geom}{the ggplot "tile" geom layer to be shown, default FALSE.}

\item{dot.col}{dot colors, default "c("white","#990000")".}

\item{midpoint}{the midpoint value when using rescale parameters, default 0.5.}

\item{scale}{whether scale the gene expressions, default TRUE.}

\item{col.min}{minimum scaled average expression threshold (everything smaller will be set to this), default -2.5.}

\item{col.max}{maximum scaled average expression threshold (everything larger will be set to this), default 2.5.}

\item{rescale}{whether rescale the average expression to specific range, default FALSE.}

\item{rescale.min}{minimum rescaled average expression threshold, default 0.}

\item{rescale.max}{maximum rescaled average expression threshold, default 1.}

\item{dot.min}{the minimum size of dot point, default 1.}

\item{dot.max}{the maximum size of dot point, default 6.}

\item{base_size}{the theme base size, default 14.}

\item{x.text.angle}{the x text angle, default 90.}

\item{x.text.hjust}{the x text hjust, default 1}

\item{x.text.vjust}{the x text vjust, default 0.}

\item{ytree}{whether add the y axis dendrogram, default TRUE.}

\item{xtree}{whether add the x axis dendrogram, default FALSE.}

\item{tree.pos}{the dendrogram position to be placed, default NULL.}

\item{same.pos.label}{whether place the x/y axis label on same side, default FALSE.}

\item{size.width}{the point legend width, default 0.3.}

\item{bar.width}{the colorbar legend width, default 4.5.}

\item{plot.margin}{the plot margin, default c(1,1,1,1) which relative to c(t,r,b,l).}

\item{anno}{whether anno celltype, default FALSE.}

\item{aesGroName}{the markerGene data.frame column name which refers to celltype, default "cluster".}

\item{segWidth}{annoSegment width, default 0.8.}

\item{lwd}{annoSegment line size, default 3.}

\item{textRot}{annoSegment text angle, default 90.}

\item{textSize}{annoSegment text size, default 14.}

\item{hjust}{annoSegment text hjust, default 0.}

\item{legend.position}{ggplot legend position, default "right".}

\item{bar.legendTitle}{colorbar legend title, default "Mean expression in group".}

\item{point.lengdTitle}{point size legend title, default "Fraction of cells in group (\%)".}

\item{...}{other parameters passed to annoSegment function.}

\item{gene.order}{supply your own gene orders, default NULL.}

\item{cluster.order}{supply your own cluster number orders, default NULL.}
}
\value{
Return a ggplot object.
}
\description{
using dotplot to visualize gene expression
}
\examples{
httest <- system.file("extdata", "htdata.RDS", package = "scRNAtoolVis")
pbmc <- readRDS(httest)

# add groups
pbmc$groups <- rep(c('stim','control'),each = 1319)
# add celltype
pbmc$celltype <- Seurat::Idents(pbmc)

# load markergene
markergene <- data("top3pbmc.markers")

# ====================================
jjDotPlot(object = pbmc,
          gene = top3pbmc.markers$gene)

jjDotPlot(object = pbmc,
          gene = top3pbmc.markers$gene,
          id = 'celltype')

jjDotPlot(object = pbmc,
          markerGene = top3pbmc.markers)

jjDotPlot(object = pbmc,
          markerGene = top3pbmc.markers,
          xtree = TRUE)

jjDotPlot(object = pbmc,
          markerGene = top3pbmc.markers,
          anno = TRUE,
          plot.margin = c(3,1,1,1))
}
\author{
Junjun Lao
}
