% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjVolcano.R
\name{jjVolcano}
\alias{jjVolcano}
\title{using jjVolcano to visualize marker genes}
\arguments{
\item{diffData}{diff results with data.frame format, default NULL.}

\item{myMarkers}{whether supply your own gene labels, default NULL.}

\item{log2FC.cutoff}{log2FoldChange cutoff, default 0.25.}

\item{pvalue.cutoff}{pvalue cutoff to filter, default 0.05.}

\item{adjustP.cutoff}{adjusted pvalue cutoff to be colored in plot, default 0.01.}

\item{topGeneN}{top genes to be labeled in plot, default 5.}

\item{col.type}{point color type("updown/adjustP"), default "updown".}

\item{back.col}{background color, default "grey93".}

\item{pSize}{point size, default 0.75.}

\item{aesCol}{point mapping color, default c("#0099CC","#CC3333").}

\item{legend.position}{legend position in plot, default c(0.7,0.9).}

\item{base_size}{theme base size, default 14.}

\item{tile.col}{cluster tile fill color, default jjAnno::useMyCol("paired",n = 9).}

\item{...}{other arguments passed by "geom_text_repel".}

\item{cluster.order}{whether given your cluster orders, default NULL.}

\item{polar}{whether make the plot to br polar, default FALSE.}

\item{expand}{the y axis expand, default c(-1,1).}

\item{flip}{whether flip the plot, default FALSE.}

\item{order.by}{top marker gene selection method, how the order is, default c("avg_log2FC").}
}
\value{
a ggplot object.
}
\description{
using jjVolcano to visualize marker genes
}
\examples{
\dontrun{
jjVolcano(diffData = pbmc.markers)
}
}
\author{
Junjun Lao
}
