% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerVocalno.R
\name{markerVocalno}
\alias{markerVocalno}
\title{Marker genes vocalno plot}
\arguments{
\item{markers}{Dataframe marker genes from findAllmarkers function from seurat.}

\item{topn}{Numbers top genes to label, defaults is 5.}

\item{log2FC}{The threshold of log2FC, defaults is 0.25.}

\item{hlineSize}{Hline size, defaults is 1.}

\item{hlineColor}{Hline color, defaults is 'grey50'.}

\item{pforce}{Positive gene force parameters to avoid overlap gene labels, defaults is 5.}

\item{nforce}{Negtive gene force parameters to avoid overlap gene labels, defaults is 2.5.}

\item{nudge_x}{Ajustments on the horizotal of the gene label, defaults is 0.8.}

\item{pnudge_y}{Ajustments on the horizotal of the positive gene label, defaults is 0.25.}

\item{nnudge_y}{Ajustments on the horizotal of the negtive gene label, defaults is 0.}

\item{base_size}{Theme base size, defaults is 14.}

\item{facetColor}{Facet border color, defaults is NA.}

\item{facetFill}{Facet fill color, defaults is 'white'.}

\item{ylab}{Plot y label, defaults is 'Log2-Fold Change'.}

\item{nrow}{Numbers rows to plot, defaults is 1.}
}
\value{
Return a ggplot.
}
\description{
Marker genes vocalno plot
}
\examples{
test <- system.file("extdata", "pbmc.markers.csv", package = "scRNAtoolVis")
markers <- read.csv(test)

markerVocalno(markers = markers,
              topn = 5,
              labelCol = ggsci::pal_npg()(9))
}
\author{
Junjun Lao
}
