% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerVolcano.R
\name{markerVolcano}
\alias{markerVolcano}
\title{Marker 基因“火山分面图”（按 cluster 分面）}
\arguments{
\item{markers}{数据框；marker 结果。至少需要列：\code{cluster}, \code{gene}, \code{avg_log2FC}, \code{pct.1}, \code{pct.2}。}

\item{ownGene}{字符向量；若提供，则只标注这些基因（覆盖 topn 逻辑），默认 NULL。}

\item{topn}{整数；每个 cluster 标注的 top 基因数量（正/负各 topn），默认 5。}

\item{log2FC}{数值；横线阈值（±log2FC），默认 0.25。}

\item{hlineSize}{Hline size, defaults is 1.}

\item{hlineColor}{Hline color, defaults is "grey50".}

\item{pforce}{Positive gene force parameters to avoid overlap gene labels, defaults is 5.}

\item{nforce}{Negative gene force parameters to avoid overlap gene labels, defaults is 2.5.}

\item{nudge_x}{Adjustments on the horizontal of the gene label, defaults is 0.8.}

\item{pnudge_y}{Adjustments on the horizontal of the positive gene label, defaults is 0.25.}

\item{nnudge_y}{Adjustments on the horizontal of the negative gene label, defaults is 0.}

\item{base_size}{Theme base size, defaults is 14.}

\item{facetColor}{Facet border color, defaults is NA.}

\item{facetFill}{Facet fill color, defaults is "white".}

\item{ylab}{Plot y label, defaults is "Log2-Fold Change".}

\item{nrow}{Numbers rows to plot, defaults is 1.}
}
\value{
Return a ggplot.
}
\description{
输入 marker 结果，绘制（pct.1 - pct.2）vs avg_log2FC 的分面图，并自动标注 top 基因。
}
\examples{
test <- system.file("extdata", "pbmc.markers.csv", package = "scRNAtoolVis")
markers <- read.csv(test)

markerVolcano(
  markers = markers,
  topn = 5,
  labelCol = ggsci::pal_npg()(9)
)
}
\author{
Junjun Lao
}
