% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjVolcano.R
\docType{data}
\name{pbmc.markers}
\alias{pbmc.markers}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 5921 rows and 7 columns.
}
\usage{
pbmc.markers
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
