% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterCellPlot.R
\name{scatterCellPlot}
\alias{scatterCellPlot}
\title{Scatter Cell Plot}
\usage{
scatterCellPlot(
  object = NULL,
  color = NULL,
  dim = "umap",
  rm.axis = FALSE,
  cell.id = NULL,
  bar.width = 0.2,
  point.size = 1,
  rm.barplot = FALSE,
  legend.psize = 1.5,
  arrow.len = 0.2
)
}
\arguments{
\item{object}{A Seurat object containing the data.}

\item{color}{A vector of colors for each cell type. If NULL, random colors will be assigned.}

\item{dim}{The dimension used for plotting (default is "umap").}

\item{rm.axis}{Logical value indicating whether to remove the x and y-axis (default is FALSE).}

\item{cell.id}{Name of the column in the metadata that represents cell identity (default is NULL).}

\item{bar.width}{Width of the barplot (default is 0.2).}

\item{point.size}{Size of the points in the scatter plot (default is 1).}

\item{rm.barplot}{whether remove barplot, default FALSE.}

\item{legend.psize}{legend point size, default 1.5.}

\item{arrow.len}{arrow length, default 0.2.}
}
\value{
None
}
\description{
This function creates a scatter cell plot using the grid package in R.
}
\examples{
\dontrun{scatterCellPlot(object = seurat_object)}

}
\author{
Jun Zhang
}
