% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjDotPlot.R
\docType{data}
\name{top3pbmc.markers}
\alias{top3pbmc.markers}
\title{This is a test data for this package
test data description}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 27 rows and 7 columns.
}
\usage{
top3pbmc.markers
}
\description{
This is a test data for this package
test data description
}
\author{
Junjun Lao
}
\keyword{datasets}
