% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracksPlot.R
\name{tracksPlot}
\alias{tracksPlot}
\title{Generate a track or heatmap plot}
\usage{
tracksPlot(
  object = NULL,
  plot.type = c("track", "heatmap"),
  genes = NULL,
  vmin = -2,
  vmax = 2,
  cell.order = NULL,
  gene.order = NULL,
  facet_nested_params = list(),
  theme_params = list(),
  strip_nested_params = list()
)
}
\arguments{
\item{object}{An optional object containing the data.}

\item{plot.type}{The type of plot to generate, either "track" or "heatmap".}

\item{genes}{A vector or data frame specifying the genes to plot.}

\item{vmin}{The minimum value for the color scale (only applicable for the heatmap plot).}

\item{vmax}{The maximum value for the color scale (only applicable for the heatmap plot).}

\item{cell.order}{An optional vector specifying the order of cells in the plot.}

\item{gene.order}{An optional vector specifying the order of genes in the plot.}

\item{facet_nested_params}{A list of additional parameters to customize the facet_nested plot.}

\item{theme_params}{A list of additional parameters to customize the plot's theme.}

\item{strip_nested_params}{A list of additional parameters to customize the strip_nested plot.}
}
\value{
A ggplot object representing the track or heatmap plot.
}
\description{
This function generates a track or heatmap plot based on the provided data.
}
\author{
Jun Zhang
}
