% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageHeatmap.R
\name{AverageHeatmap}
\alias{AverageHeatmap}
\title{Plot averaged gene expression cross cluster cells}
\usage{
AverageHeatmap(
  object = NULL,
  markerGene = NULL,
  group.by = "ident",
  assays = "RNA",
  slot = "data",
  htCol = c("#0099CC", "white", "#CC0033"),
  colseed = 666,
  htRange = c(-2, 0, 2),
  annoCol = FALSE,
  myanCol = NULL,
  annoColType = "light",
  annoColTypeAlpha = 0,
  row_title = "Cluster top Marker genes",
  clusterAnnoName = TRUE,
  showRowNames = TRUE,
  row_names_side = "left",
  markGenes = NULL,
  border = FALSE,
  fontsize = 10,
  column_names_rot = 45,
  width = NULL,
  height = NULL,
  cluster.order = NULL,
  cluster_columns = FALSE,
  cluster_rows = FALSE,
  ...
)
}
\arguments{
\item{object}{object seurat object.}

\item{markerGene}{Your marker genes.}

\item{group.by}{Categories for grouping (e.g, ident, replicate, celltype). 'ident' by default.}

\item{assays}{Which assays to use. Default is "RNA" assays.}

\item{slot}{Slot(s) to use. Default is "data"}

\item{htCol}{Heatmap colors. Default is c("#0099CC", "white", "#CC0033").}

\item{colseed}{Cluster annotaion colors seed, these colors are produed randomly, so you can give a seed to assure produce same colors.  Default is 666.}

\item{htRange}{Heatmap values range. Default is c(-2, 0, 2).}

\item{annoCol}{Whether use your own annotation clusters colors. Default is 'FALSE'.}

\item{myanCol}{You can specify your own annotation clusters colors vectors. Default is 'null'.}

\item{annoColType}{Cluster annotaion colors type (bright, light, dark and random). Default is light.}

\item{annoColTypeAlpha}{Cluster annotaion colors transparency. Default is 0.}

\item{row_title}{Heatmap row title. Default is "Cluster top Marker genes".}

\item{clusterAnnoName}{Whether to add clsuetr column annotation name. Default is "TRUE".}

\item{showRowNames}{whether to show rownames. Default is "TRUE".}

\item{row_names_side}{Heatmap gene name side. Default is "left".}

\item{markGenes}{Provide your tartget genes to mark on the plot. Default is "NULL".}

\item{border}{Whether to shOw heatmap border. Default is "FALSE".}

\item{fontsize}{Heatmap gene name fontsize. Default is 10.}

\item{column_names_rot}{Cluster name rotation. Default is 45.}

\item{width}{The heatmap body width. Default is "NULL".}

\item{height}{The heatmap body height. Default is "NULL".}

\item{cluster.order}{The cell clusters order. Default is "NULL".}

\item{cluster_columns}{Whether cluster columns. Default is "FALSE".}

\item{cluster_rows}{Whether cluster rows. Default is "FALSE".}

\item{...}{Other arguments passed with ComplexHeatmap::rowAnnotation and ComplexHeatmap::Heatmap.}
}
\value{
Return a plot.
}
\description{
Plot averaged gene expression cross cluster cells
}
\examples{
httest <- system.file("extdata", "htdata.RDS", package = "scRNAtoolVis")
pbmc <- readRDS(httest)

# load markergene
markergene <- system.file("extdata", "top5pbmc.markers.csv", package = "scRNAtoolVis")
markers <- read.table(markergene, sep = ",", header = TRUE)

# plot
AverageHeatmap(
  object = pbmc,
  markerGene = markers$gene
)

# change color
AverageHeatmap(
  object = pbmc,
  markerGene = markers$gene,
  htCol = c("#339933", "#FFCC00", "#FF0033")
)

}
\author{
Junjun Lao
}
