% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCornerAxes.R
\name{clusterCornerAxes}
\alias{clusterCornerAxes}
\title{在 UMAP/tSNE 聚类图上添加“角标坐标轴”（Corner Axes）}
\arguments{
\item{object}{Seurat 对象，必须包含对应降维（例如 "umap"）。}

\item{reduction}{字符串；降维名称（umap/tsne 等），默认 "umap"。}

\item{groupFacet}{字符串；当 \code{noSplit = FALSE} 时用于分面（facet）的 meta.data 列名。}

\item{clusterCol}{"string", the point color to group by,cluster name, default "seurat_clusters".}

\item{pSize}{"num", point size.}

\item{aspect.ratio}{"num", plot width and height ratio, default NULL.}

\item{noSplit}{"logic", whether to split/facet the plot, default "TRUE".}

\item{nrow}{"num", rows to plot when noSplit = FALSE.}

\item{relLength}{"num", the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num" ,the relative distance of corner axis label to axis.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), default "mul".}

\item{legendPos}{"string", legend position same as ggplot theme function, default "right".}

\item{keySize}{The legned point size, default is 5.}

\item{lineTextcol}{"string", corner line and label color, default "black".}

\item{stripCol}{"string", facet background color, default "white".}

\item{arrowType}{"string", arrow type (open/closed), default "closed".}

\item{cornerTextSize}{"num", the corner label text size, default is 3.}

\item{base_size}{"num", theme base size, default is 14.}

\item{themebg}{Another theme style, default is "default", or "bwCorner".}

\item{addCircle}{"logic", whether add circle on clusters, default is "FALSE".}

\item{addCircle.legacy}{"logic", using the legacy version to add a circle, the parameters \code{nbin}, \code{nsm}, \code{addsm}, \code{sfac} and \code{qval} are only applicable to legacy, default is "FALSE".}

\item{cicAlpha}{"num", circle fill color alpha, default is 0.1.}

\item{cicDelta}{"num", the distance to extend the curve (circle), this parameter only takes effect when \code{addCircle.legacy = FALSE}. 默认 0.1。}

\item{cicLineSize}{"num", circle line size, default is 1.}

\item{cicLineColor}{"num", circle line color, default is "grey50".}

\item{cicLineLty}{"num", circle line type, default is "dashed".}

\item{nbin}{"num", number of points used to shape the hull, default 100.}

\item{nsm}{"num", number of points used to perform convolution, should less than nbin, default 10.}

\item{addsm}{"num", number of additional times of convolution performed, default 1.}

\item{sfac}{"num", quantile of each sector, used to determine the edge of the hull, should less than 1, default 1.}

\item{qval}{"num", expansion size factor, larger value means bigger hull, default 1.5.}

\item{cellLabel}{Whether to label cell type on plot, default is FALSE.}

\item{cellLabelSize}{Cell type label size, default is 6.}

\item{cellLabelColor}{Cell type label color, default is "black".}

\item{show.legend}{Whether show legend, default is TRUE.}
}
\value{
返回一个 ggplot 对象。
}
\description{
绘制聚类散点图（按 cluster 上色），并在左下角添加角标坐标轴；可选为每个 cluster 画外轮廓圈。
}
\examples{
test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = TRUE
)

# arrowType
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = TRUE, arrowType = "open"
)

# facet by metadata column "orig.ident"
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5
)

# retain only one axes
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5,
  axes = "one"
)

# line color
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5,
  lineTextcol = "grey50"
)

# tsne
clusterCornerAxes(
  object = tmp, reduction = "tsne",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5
)

}
\author{
Junjun Lao
}
