% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterCornerAxes.R
\name{clusterCornerAxes}
\alias{clusterCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE cluster figures}
\arguments{
\item{object}{seurat object.}

\item{reduction}{"string", reduction type (umap/tsne).}

\item{groupFacet}{"string", give the column name in seurat metadata to facet plot.}

\item{clusterCol}{"string", the point color to group by,cluster name, defaults "seurat_clusters".}

\item{pSize}{"num", point size.}

\item{noSplit}{'logic', whether to split/facet the plot, defaults "TRUE".}

\item{nrow}{"num", rows to plot when noSplit = FALSE.}

\item{relLength}{'num', the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num" ,the relative distance of corner axis label to axis.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), defaults "mul".}

\item{legendPos}{"string", legend position same as ggplot theme function, defaults "right".}

\item{lineTextcol}{"string", corner line and label color, defaults "black".}

\item{stripCol}{"string", facet balckground color, defaults "white".}

\item{arrowType}{"string", arrow type (open/closed), defaults "closed".}

\item{cornerTextSize}{"num", the corner label text size, defaults is 5.}

\item{base_size}{"num", theme base size, defaults is 14.}

\item{themebg}{Another theme style, defaults is 'default', or 'bwCorner'.}

\item{addCircle}{Logic, whether add circle on clusters, default is 'FALSE'.}

\item{cicAlpha}{"num", circle fill color alpha, default is 0.1,}

\item{cicLineSize}{"num", circle line size, default is 1.}

\item{cicLineColor}{"num", circle line color, default is 'grey50'.}

\item{cicLineLty}{"num", circle line type, default is 'dashed'.}

\item{nbin}{"num", number of points used to shape the hull, default 100.}

\item{nsm}{"num", number of points used to perform convolution, should less than nbin, default 10.}

\item{addsm}{"num", number of additional times of convolution performed, default 1.}

\item{sfac}{"num", quantile of each sector, used to determine the edge of the hull, should less than 1, default 1.}

\item{qval}{"num", expansion size factor, larger value means bigger hull, default 1.5.}
}
\value{
Return a ggplot object.
}
\description{
Add corner axes on seurat UMAP/tSNE cluster figures
}
\examples{
test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = TRUE
)

# arrowType
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = TRUE, arrowType = "open"
)

# facet by metadata column "orig.ident"
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5
)

# retain only one axes
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5,
  axes = "one"
)

# line color
clusterCornerAxes(
  object = tmp, reduction = "umap",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5,
  lineTextcol = "grey50"
)

# tsne
clusterCornerAxes(
  object = tmp, reduction = "tsne",
  noSplit = FALSE, groupFacet = "orig.ident",
  relLength = 0.5
)

}
