% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCornerAxes.R
\name{featureCornerAxes}
\alias{featureCornerAxes}
\title{给 FeaturePlot 加“角标坐标轴”（Corner Axes）}
\arguments{
\item{object}{Seurat 对象。}

\item{reduction}{字符串；降维名称（umap/tsne），默认 "umap"。}

\item{features}{字符向量；要绘制的基因列表。}

\item{slot}{读取表达矩阵的层（兼容 Seurat v4/v5），默认 "data"。
\itemize{
\item Seurat v4：等价于 \code{FetchData(..., slot = slot)}
\item Seurat v5：等价于 \code{FetchData(..., layer = slot)}
}}

\item{groupFacet}{字符串；用于分面的 meta.data 列名。若为 NULL，则只按基因分面。}

\item{relLength}{"num", the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num", the relative distance of corner axis label to axis.}

\item{aspect.ratio}{"num", plot width and height ratio, default NULL.}

\item{low}{"string", point color with low expression.}

\item{high}{"string", point color with high expression.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), default "mul".}

\item{legendPos}{"string", legend position same as ggplot theme function, default "right".}

\item{stripCol}{"string", facet background color, defaults "white".}

\item{pSize}{"num", point size.}

\item{arrowType}{"string", arrow type (open/closed), default "closed".}

\item{lineTextcol}{"string", facet background color, default "white".}

\item{cornerTextSize}{"num", the corner label text size, default is 5.}

\item{base_size}{"num", theme base size, default is 14.}

\item{themebg}{Another theme style, default is "default", or "bwCorner".}

\item{show.legend}{Whether show legend, default "TRUE".}

\item{cornerVariable}{Which group  corner axis to be added when "axes" set to "one", default is the first group.}

\item{nLayout}{= NULL Similar to the ncol/nrow for the layout, default is the gene numbers.}

\item{minExp}{Minimum expression value defined, default is NULL.}

\item{maxExp}{Maxmum expression value defined, default is NULL.}
}
\value{
Return a ggplot.
}
\description{
基于 Seurat 的降维坐标与基因表达绘制散点图，并在左下角添加“角标坐标轴”（Corner Axes）。
}
\examples{

test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
featureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2"),
  slot = "data"
)

# one axes
featureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  features = c("Actb", "Ythdc1", "Ythdf2"),
  relLength = 0.5, relDist = 0.2,
  axes = "one",
  lineTextcol = "grey50",
  slot = "data"
)

# tsne
featureCornerAxes(
  object = tmp, reduction = "tsne",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2"),
  slot = "data"
)


}
\author{
Junjun Lao
}
