% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureCornerAxes.R
\name{featureCornerAxes}
\alias{featureCornerAxes}
\title{Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures}
\arguments{
\item{object}{object seurat object.}

\item{reduction}{"string", reduction type (umap/tsne).}

\item{features}{"string", the gene you want to plot.}

\item{groupFacet}{"string", give the column name in seurat metadata to facet plot, if it is "NULL", facet plot only by gene.}

\item{relLength}{"num", the corner axis line relative length to plot axis(0-1).}

\item{relDist}{"num", the relative distance of corner axis label to axis.}

\item{aspect.ratio}{"num", plot width and height ratio, default NULL.}

\item{low}{"string", point color with low expression.}

\item{high}{"string", point color with high expression.}

\item{axes}{"string", show multiple corner axis or only one (mul/one), default "mul".}

\item{legendPos}{"string", legend position same as ggplot theme function, default "right".}

\item{stripCol}{"string", facet background color, defaults "white".}

\item{pSize}{"num", point size.}

\item{arrowType}{"string", arrow type (open/closed), default "closed".}

\item{lineTextcol}{"string", facet background color, default "white".}

\item{cornerTextSize}{"num", the corner label text size, default is 5.}

\item{base_size}{"num", theme base size, default is 14.}

\item{themebg}{Another theme style, default is "default", or "bwCorner".}

\item{show.legend}{Whether show legend, default "TRUE".}

\item{cornerVariable}{Which group  corner axis to be added when "axes" set to "one", default is the first group.}

\item{nLayout}{= NULL Similar to the ncol/nrow for the layout, default is the gene numbers.}

\item{minExp}{Minimum expression value defined, default is NULL.}

\item{maxExp}{Maxmum expression value defined, default is NULL.}
}
\value{
Return a ggplot.
}
\description{
Add corner axes on seurat UMAP/tSNE gene FeaturePlot function figures
}
\examples{

test <- system.file("extdata", "seuratTest.RDS", package = "scRNAtoolVis")

tmp <- readRDS(test)

# umap
featureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)

# one axes
featureCornerAxes(
  object = tmp, reduction = "umap",
  groupFacet = "orig.ident",
  features = c("Actb", "Ythdc1", "Ythdf2"),
  relLength = 0.5, relDist = 0.2,
  axes = "one",
  lineTextcol = "grey50"
)

# tsne
featureCornerAxes(
  object = tmp, reduction = "tsne",
  groupFacet = "orig.ident",
  relLength = 0.5, relDist = 0.2,
  features = c("Actb", "Ythdc1", "Ythdf2")
)


}
\author{
Junjun Lao
}
